Move MockedPulsarStandalone to other package
